requirements.txt
flask
flask-cors
pandas
numpy
google-generativeai
python-dotenv
matplotlib
seaborn
